First commit
First commmit to the repo. A jupyter notebook is uploaded that uses geopandas in order to plot trajectories in their correspondent map. This is done to analyse if there are 1-to-1 relationships between the real and synthetic trajectories that have been generated with Differential Privacy. Further exploration needs to be performed